fixes issue #5.
needed to do another intersection test for the virtual trace
after magnetization.  Also needed to update the source and destination
hit list.  This should be made more modular so that we can re-use
it for other components.  This issue might exist in other areas
of magnetization placement but have been missed because it's not
as common as the pad magnet issue.

I think the reason why it was mostly only happening on initial
joint pair instead of having middle tracks was because that
condition happens to be more constrained than coming at it from
a better angle.

In one of the tests I was doing, the 'dst' hit list included
another pad so we thought we were placing a trace that connected
both of those pads.  After magnetization, the trace stopped intersecting
the pad but was still in the critical clearance region.  Since
the 'dst' hit list hadn't been updated after magnetization and it
was still using the old dst list, the intersection passed and it placed.

That's why this issue might still be present for other combinations.
We'll keep an eye out for it but the pad issue was the most visible.