fiddling with compent_lib text field
The component would be fine but the text would be
displayed incorrectly in the footprint tile library
section and would take on the value.  load_part
updates the component_lib with the most current data.
That's fine (I think) but when displaying we really
need to know if we're displaying a footprint
from the board or a library element.  If we're
displaying a library element, show the library_name,
not the text field name.